+++ src/common/WebComponent.ts
@@ -1,6 +1,19 @@
 interface WebComponent {
-  render(): string;
+  render(): Element;
   setup(): void;
 }
 
-export { WebComponent };
+function createElementFromString(elementString: string): Element {
+  const element = document.createElement('div');
+  element.innerHTML = elementString;
+
+  const createdElement = element.firstElementChild;
+  if (createdElement === null) {
+    throw 'Incorrect argument';
+  }
+
+  return createdElement;
+}
+
+export default WebComponent;
+export { createElementFromString };