+++ src/components/sample/sample.spec.ts
@@ -1,5 +1,0 @@
-import { SampleComponent } from './sample';
-
-test('Test rendering', () => {
-  expect((new SampleComponent()).render()).toBe('<p class="my-p" id="myP">sample text</p>');
-});