+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - HTML Validator.
 - Web Component interface.
 - Images for ui.
+- Router class.
 
 ## [0.1.0] - 2021-01-25